tests/wpt/web-platform-tests/css/filter-effects/parsing/color-interpolation-filters-parsing-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Filter Effects Module Level 1: parsing color-interpolation-filters with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/filter-effects/#ColorInterpolationFiltersProperty">
<link rel="help" href="https://drafts.csswg.org/cssom/#serializing-css-values">
<meta name="assert" content="color-interpolation-filters supports the full grammar 'auto | sRGB | linearRGB'.">
<meta name="assert" content="Keywords serialize in lower case.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("color-interpolation-filters", "auto");
test_valid_value("color-interpolation-filters", "sRGB", "srgb");
test_valid_value("color-interpolation-filters", "linearRGB", "linearrgb");
test_valid_value("color-interpolation-filters", "LiNeArRgB", "linearrgb");
</script>
</body>
</html>
